Add files via upload
Hangman is a word-guessing game where one player thinks of a word and the other player tries to guess it by suggesting letters. In the Hangman project, the goal is to create a computer program that allows users to play the game against the computer. The program randomly selects a word, and the user tries to guess the word by suggesting letters while the program displays a hangman figure for incorrect guesses.